Added users to image addition/removal.

+++ Shopify Challenge/src/imageRepository/ImageRepository.java
@@ -28,6 +28,7 @@ import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
@@ -44,6 +45,7 @@ import javax.imageio.ImageIO;
  */
 public final class ImageRepository {
 	private static final String USERS_FILE_PATH = "users.txt";
+	private static final String IMAGEDATA_FILE_PATH = "imagedata.txt";
 	
 	/**
 	 * Loads an image repository from a directory {@code imageDir}.
@@ -51,14 +53,51 @@ public final class ImageRepository {
 	 * @since 2021-01-17
 	 */
 	public static final ImageRepository fromDirectory(File imageDir) {
+		final Map<String, ImageEntry> data = loadImageData(
+				new File(imageDir, IMAGEDATA_FILE_PATH));
+		final List<User> users = loadUsers(new File(imageDir, USERS_FILE_PATH));
+		
+		// read imagedata.txt to get data on images
+		return new ImageRepository(imageDir, data, users);
+	}
+	
+	/**
+	 * Loads image data from the file {@code file}.
+	 *
+	 * @since 2021-01-17
+	 */
+	private static final Map<String, ImageEntry> loadImageData(File file) {
 		final Map<String, ImageEntry> data = new HashMap<>();
-		for (final File f : imageDir.listFiles()) {
-			if (!USERS_FILE_PATH.equals(f.getName())) {
-				data.put(f.getName(), ImageEntry.loadImage(f.getName()));
+		
+		try (final BufferedReader reader = new BufferedReader(
+				new FileReader(file))) {
+			String line = null;
+			while ((line = reader.readLine()) != null) {
+				final String[] parts = line.split(":");
+				if (parts.length != 3)
+					throw new IllegalStateException("Invalid imagedata.txt file.");
+				
+				// split get data
+				final String name = parts[0];
+				final String username = parts[1];
+				final boolean isPublic = "public".equals(parts[2]);
+				
+				// convert data to entry
+				final ImageEntry entry;
+				if (username == "") {
+					entry = ImageEntry.loadImage(name);
+				} else {
+					entry = ImageEntry.loadImage(name, username, isPublic);
+				}
+				
+				// insert entry into database
+				data.put(name, entry);
 			}
+		} catch (final IOException e) {
+			e.printStackTrace();
 		}
-		return new ImageRepository(imageDir, data,
-				loadUsers(new File(imageDir, USERS_FILE_PATH)));
+		
+		return data;
 	}
 	
 	/**
@@ -111,16 +150,22 @@ public final class ImageRepository {
 	 * @param originalFilepath place where image was originally stored
 	 * @param newFilepath      place where the image will be stored, relative to
 	 *                         directory, must be in repository directory.
+	 * @param username         username of user who added the image
+	 * @param isPublic         true iff the image is public
 	 * @since 2021-01-17
 	 */
-	public final void addImage(File originalFilepath, String newFilepath) {
-		this.data.put(newFilepath, ImageEntry.loadImage(newFilepath));
+	public final void addImage(File originalFilepath, String newFilepath,
+			String username, boolean isPublic) {
+		this.data.put(newFilepath,
+				ImageEntry.loadImage(newFilepath, username, isPublic));
 		try {
 			Files.copy(Path.of(originalFilepath.getAbsolutePath()),
 					this.getPath(newFilepath));
 		} catch (final IOException e) {
 			e.printStackTrace();
 		}
+		
+		this.saveImageData();
 	}
 	
 	/**
@@ -154,6 +199,15 @@ public final class ImageRepository {
 		}
 	}
 	
+	/**
+	 * Gets data about the image {@code name}. Does not load the image file.
+	 *
+	 * @since 2021-01-17
+	 */
+	public final ImageEntry getImageData(String name) {
+		return this.data.get(name);
+	}
+	
 	private final Path getPath(String imageFilename) {
 		return Path.of(this.directory.getAbsolutePath(), imageFilename);
 	}
@@ -171,26 +225,46 @@ public final class ImageRepository {
 	}
 	
 	/**
+	 * @param username user to get list for; null for no user
 	 * @return set of names of all images in repository
 	 * @since 2021-01-17
 	 */
-	public final Set<String> imageNames() {
-		return Collections.unmodifiableSet(this.data.keySet());
+	public final Set<String> imageNames(String username) {
+		final Set<String> imageNamesFiltered = new HashSet<>(this.data.keySet());
+		imageNamesFiltered.removeIf(name -> {
+			final ImageEntry data = this.getImageData(name);
+			return !data.isPublic() && !data.getUser().get().equals(username);
+		});
+		return Collections.unmodifiableSet(imageNamesFiltered);
 	}
 	
 	/**
 	 * Removes an image from the directory
 	 *
-	 * @param name name of image to remove
+	 * @param name     name of image to remove
+	 * @param username username of remover, must match owner's username, null for
+	 *                 not logged in
+	 * 						
+	 * @return true iff the removal was successful
 	 * @since 2021-01-17
 	 */
-	public final void removeImage(String name) {
-		this.data.remove(name);
-		try {
-			Files.delete(this.getPath(name));
-		} catch (final IOException e) {
-			e.printStackTrace();
-		}
+	public final boolean removeImage(String name, String username) {
+		final ImageEntry entry = this.data.get(name);
+		
+		if (entry.getUser().isPresent()
+				&& entry.getUser().get().equals(username)) {
+			this.data.remove(name);
+			try {
+				Files.delete(this.getPath(name));
+			} catch (final IOException e) {
+				e.printStackTrace();
+			}
+			
+			this.saveImageData();
+			
+			return true;
+		} else
+			return false;
 	}
 	
 	/**
@@ -216,9 +290,26 @@ public final class ImageRepository {
 	}
 	
 	/**
+	 * Saves all image data.
+	 * 
+	 * @since 2021-01-17
+	 */
+	public final void saveImageData() {
+		final File file = new File(this.directory, IMAGEDATA_FILE_PATH);
+		
+		try (BufferedWriter writer = new BufferedWriter(new FileWriter(file))) {
+			for (final String name : this.data.keySet()) {
+				final ImageEntry entry = this.data.get(name);
+				writer.write(entry.toString() + "\n");
+			}
+		} catch (final IOException e) {
+			e.printStackTrace();
+		}
+	}
+	
+	/**
 	 * Saves the user data into the specified file.
 	 *
-	 * @param file file to save into
 	 * @since 2021-01-17
 	 */
 	public final void saveUsers() {
